feat(colpali_engine/collators): added `TAIHardNegCollator`.
The default datasets that colpali uses, have a structed directory schema, with the images existing in different directories. We do not store the images in a separate directory. So we need our own collator to read the data from the training data JSONL to read the images and convert them to image objects that can be used for fine-tuning